Refactored todos and added age during sleep
