+++ test/api/v3/unit/libs/cron.test.js
@@ -340,99 +340,6 @@ describe('cron', () => {
     });
   });
 
-  describe('todos', () => {
-    beforeEach(() => {
-      let todo = {
-        text: 'test todo',
-        type: 'todo',
-        value: 0,
-      };
-
-      let task = new Tasks.todo(Tasks.Task.sanitize(todo)); // eslint-disable-line new-cap
-      tasksByType.todos.push(task);
-    });
-
-    it('should make uncompleted todos redder', () => {
-      let valueBefore = tasksByType.todos[0].value;
-      cron({user, tasksByType, daysMissed, analytics});
-      expect(tasksByType.todos[0].value).to.be.lessThan(valueBefore);
-    });
-
-    it('should add history of completed todos to user history', () => {
-      tasksByType.todos[0].completed = true;
-
-      cron({user, tasksByType, daysMissed, analytics});
-
-      expect(user.history.todos).to.be.lengthOf(1);
-    });
-
-    it('should remove completed todos from users taskOrder list', () => {
-      tasksByType.todos = [];
-      user.tasksOrder.todos = [];
-      let todo = {
-        text: 'test todo',
-        type: 'todo',
-        value: 0,
-      };
-
-      let task = new Tasks.todo(Tasks.Task.sanitize(todo)); // eslint-disable-line new-cap
-      tasksByType.todos.push(task);
-      task = new Tasks.todo(Tasks.Task.sanitize(todo)); // eslint-disable-line new-cap
-      tasksByType.todos.push(task);
-      tasksByType.todos[0].completed = true;
-
-      user.tasksOrder.todos = tasksByType.todos.map(taskTodo => {
-        return taskTodo._id;
-      });
-      // Since ideally tasksByType should not contain completed todos, fake ids should be filtered too
-      user.tasksOrder.todos.push('00000000-0000-0000-0000-000000000000');
-
-      expect(tasksByType.todos).to.be.lengthOf(2);
-      expect(user.tasksOrder.todos).to.be.lengthOf(3);
-
-      cron({user, tasksByType, daysMissed, analytics});
-
-      // user.tasksOrder.todos should be filtered while tasks by type remains unchanged
-      expect(tasksByType.todos).to.be.lengthOf(2);
-      expect(user.tasksOrder.todos).to.be.lengthOf(1);
-    });
-
-    it('should preserve todos order in task list', () => {
-      tasksByType.todos = [];
-      user.tasksOrder.todos = [];
-      let todo = {
-        text: 'test todo',
-        type: 'todo',
-        value: 0,
-      };
-
-      let task = new Tasks.todo(Tasks.Task.sanitize(todo)); // eslint-disable-line new-cap
-      tasksByType.todos.push(task);
-      task = new Tasks.todo(Tasks.Task.sanitize(todo)); // eslint-disable-line new-cap
-      tasksByType.todos.push(task);
-      task = new Tasks.todo(Tasks.Task.sanitize(todo)); // eslint-disable-line new-cap
-      tasksByType.todos.push(task);
-
-      // Set up user.tasksOrder list in a specific order
-      user.tasksOrder.todos = tasksByType.todos.map(todoTask => {
-        return todoTask._id;
-      }).reverse();
-      let original = user.tasksOrder.todos; // Preserve the original order
-
-      cron({user, tasksByType, daysMissed, analytics});
-
-      let listsAreEqual = true;
-      user.tasksOrder.todos.forEach((taskId, index) => {
-        if (original[index]._id !== taskId) {
-          listsAreEqual = false;
-        }
-      });
-
-      expect(listsAreEqual);
-      expect(user.tasksOrder.todos).to.be.lengthOf(original.length);
-    });
-  });
-
   describe('dailys', () => {
     beforeEach(() => {
       let daily = {
